Ahora la cantidad de los insumos recepcionados se agragan al stock. Configure un decoradtors.py en usuarios, para manipular los permisos sin necesidad de otorgar permisos desde administrador de django

+++ ventas/models.py
@@ -66,5 +66,5 @@ class Venta(models.Model):
 
 class Cliente_Mayorista(models.Model):
     razon_social = models.CharField(max_length=30)
-    cuit = models.PositiveSmallIntegerField()
-    telefono = models.PositiveSmallIntegerField()
+    cuit = models.PositiveIntegerField()
+    telefono = models.PositiveIntegerField()
